CORE-261: Rework BREthereumTransfer Identifier
Rename `transferGetHash` to `transferGetIdentifier` and define the identifier only to be non-empty when the transfer’s log/transaction has been included.  When a log/transfer is announced (by BCS or by the BRD backend) try to find the corresponding transfer using the ‘identifier’ and then, if the identifier doesn’t exist, with the ‘originatingTransaction’s hash.  Make the basis handling between log and transaction consistent - specifically only assign the transaction when included, which is how the assignment of log is required to be implemented.  To confirm LES ‘submit’ functionality, update BRCrypto.  For `transferGetFee()` use the ‘basis’ if it exists, otherwise use the transfer’s originatingTransaction, if it exists; otherwise fee is zero.