CORE-290: In ETH+BTC callbacks, replace `this` with `manager`.  Define `crypto/pending`
ETH and BTC callbacks used to be defined in the WalletManger but are now defined in System; the callbacks used `this` to indicate `manager` - replace `this` with `manager`.  Also deleted dead code in WalletManager related to callbacks.  Move `crypto/BRCrypto{CurrencyPair,Transfer,Wallet,WalletManager,System}` into `crypto/pending` (as those file are unused (implementation is in Swift, not C at the moment).