CORE-313: Support BlockchainFee.  Add BRChainParamsGetCheckpointBefore
In a ‘/blockchains’ DB request, parse the fee. When merging ‘/blockchains’ and ‘/currencies’ defaults be sure to merge ‘defaults into query result’ (not the other way around).  Temporarily set the DemoApp earliestKeyTime to be before the current BlockchainDB height (for BTC). In BTC/BCH BRD sync modes, derive the begBlockNumber from the checkpoint directly - instead of from the PeerManager’s last block which has been intentionally set to ‘infinity’ so as to disable P2P sync! In BRWalletManager initialize `requestId` (oops) and clean up the `bwmPeriodicDispatcher()` logic for `bwmUpdateTransactions()`.