CORE-291: For ETH, on `cryptoWalletManagerCreate()` query for ERC20 tokens
Query based on `ewmUpdateTokens()` and then as each token is created also create a corresponding EWM wallet (if the BlockChainDB has a currency for token).  If an EWM wallet is created, it will ‘event flow’ to create a CRYPTO wallet.  Take some care for the case where EWM creates a wallet based on a sync-discovered log event.  Clean up some formatting.  Fix an issue where a Swift wallet was always created with the wallet managers unit rather than the CRYPTO wallet’s unit.